extract applyDeviceQuaternion to file
